# feat: parallel `docker-compose pull`, improve `ddev debug download-images`, fixes #7163