… docs/content/users/usage/commands.md
… pkg/dockerutil/dockerutils_test.go
… pkg/util/yamltools.go