## The Issue

- #7163

## How This PR Solves The Issue

- Removes `docker pull` completely
- Uses `docker-compose pull` everywhere
- Refactors `ddev debug download-images`:
  - Make it always pull, even if the image exists
  - Add optional project arg
  - Add `--all` flag to pull everything for all projects

## Manual Testing Instructions

Speed comparison for pulling basic images (Linux AMD64):

```
ddev delete images --all --yes && docker builder prune -f
start=$(date +%s%N)
ddev debug download-images
echo "Elapsed time: $((($(date +%s%N) - start) / 1000000)) ms"
```

Before (three retries):
```
Elapsed time: 46599 ms
Elapsed time: 47083 ms
Elapsed time: 53672 ms
```

After (three retries):

```
Elapsed time: 25162 ms
Elapsed time: 24964 ms
Elapsed time: 22870 ms
```

---

Check out:

```
ddev debug download-images

cd d11 && ddev debug download-images

cd ~ && ddev debug download-images d11

ddev debug download-images --all
```

![pull](https://github.com/user-attachments/assets/025c2012-35fd-47d0-b8e4-10296a00cb9c)

## Automated Testing Overview

<!-- Please describe the tests introduced by this PR, or explain why no tests are needed. -->

## Release/Deployment Notes

<!-- Does this affect anything else or have ramifications for other code? Does anything have to be done on deployment? -->