@stasadev 2025-07-24
Regarding implementing pull on `ddev start` for *existing* images:

- https://github.com/ddev/ddev/pull/7217#issuecomment-2824369781

After thinking about it, I would prefer pulling once per day on `ddev start` rather than on every `ddev start`.

But I think this is out of scope for this PR.

---

Key change here:

- `ddev debug download-images` always pulls
- `ddev start` works as before, no pull for existing images