@rfay 2025-07-30
I got stymied by bad internet testing this. I like the beautiful new presentation, but a couple of comments:
* Instead of saying "pulled" with a checkmark, could it say the image that was pulled?
* There seems to be a leftover "lling" from the pull in the very nice presentation. (This was on a `ddev config --auto` in a project that hadn't been started)

<img width="1230" height="427" alt="image" src="https://github.com/user-attachments/assets/f59b0ca9-738a-47a9-8715-3e15b91e2f8f" />