@rfay 2025-07-30
It's looking good!

More nits... The name given isn't really the name of the pulled image. For example, we didn't pull `ddev-hobobiker-xhgui` or `ddev-d11simple-web`

```
Downloading images for all projects
[+] Pulling 45/75
[+] Pulling 73/75                                                   1.3s
 ✔ busybox Pulled                                                   1.3s
 ✔ ddev-router Pulled                                               1.1s  ✔ ddev-t3v13-db Pulled                                            23.6s
 ✔ ddev-utilities Pulled                                            1.0s
 ✔ ddev-hobobiker-xhgui Pulled                                      1.0s
 ⠋ ddev-d11simple-web [⣤] 202.6MB / 395MB   Pulling                61.1s
 ✔ ddev-d11-db Pulled                                               1.1s
 ✔ ddev-d11-redis Pulled                                           31.9s
 ✔ ddev-hobobiker-db Pulled                                        48.7s
 ✔ ddev-ssh-agent Pulled                                            1.1s
```