@rfay 2025-07-30
Did you already check this out for offline usage? I'll try to give it a shot.