@stasadev 2025-07-31
> Did you already check this out for offline usage? I'll try to give it a shot.

Yes, works as before.

> More nits... The name given isn't really the name of the pulled image. For example, we didn't pull `ddev-hobobiker-xhgui` or `ddev-d11simple-web`

That's not an image name, it's a service name. I avoided using the image name because I wasn't sure what pattern was valid there.

I found the allowed pattern here https://github.com/compose-spec/compose-go/blob/0bd910723fa2d79a3da0334cbec6f958f632936d/schema/compose-spec.json#L31

So now, each image name used as a service name during pull will look like this:

```diff
-ddev/ddev-webserver:v1.24.7
+ddev-ddev-webserver-v1.24.7

```